+++ packages/core/loglayer/src/index.ts
@@ -10,12 +10,12 @@ export type {
   ErrorOnlyOpts,
   ILogBuilder,
   ILogLayer,
+  LogLayerPlugin,
   LogLayerTransport,
   LogLevelType,
   PluginBeforeDataOutParams,
   PluginBeforeMessageOutParams,
   PluginShouldSendToLoggerParams,
-  LogLayerPlugin,
 } from "@loglayer/shared";
 export { LogLevel } from "@loglayer/shared";
 export { LogLayer } from "./LogLayer.js";